+++ Changelog.md
@@ -7,6 +7,7 @@
 - New features translated into German.
 - code refactoring - unused functions removed
 - Import dialog from menu fixed
+- Librarypage add rename file and copy filename to clipboard
 
 01.03.26, Sandy Marko Knauer <sandy.marko.k@gmail.com>